+++ lib/builtin/settings_ui/settings_ui.dart
@@ -1,2 +1,3 @@
 export 'src/extension.dart';
+export 'src/settings_category_view.dart';
 export 'src/settings_modal.dart';